http fixed and working
