+++ package.json
@@ -25,6 +25,9 @@
   },
   "homepage": "https://github.com/gSchool/fs-pet-shop#readme",
   "dependencies": {
+    "express": "^4.14.1",
+    "fs": "0.0.1-security",
+    "http-server": "^0.9.0",
     "nodemon": "^1.9.1"
   },
   "devDependencies": {
